Switched Chromebooks from mobile to 'isDesktop'
Chromebooks were flagged as being mobile and tablet, probably a mistake because of the fact they returned the 'device' field at all. This fix labels them correctly as 'isDesktop'.

Chromebooks should be considered desktop rather than mobile or tablet because of the form factor, as well as them not obeying the typical mobile/tablet rules such as not auto-playing video.